Factor out abstract surveys converter
