+++ destinations/airbyte-faros-destination/test/converters/data.ts
@@ -10,7 +10,7 @@ export const jenkinsAllStreamsLog = readTestResourceFile(
 );
 export const jiraAllStreamsLog = readTestResourceFile('jira/all-streams.log');
 export const airtableSurveysAllStreamsLog = readTestResourceFile(
-  'airtable/surveys/all-streams.log'
+  'abstract-surveys/surveys/all-streams.log'
 );
 export const asanaAllStreamsLog = readTestResourceFile('asana/all-streams.log');
 export const gitlabAllStreamsLog = readTestResourceFile(
