+++ destinations/airbyte-faros-destination/test/resources/abstract-surveys/surveys/catalog.json
